Update for split OSPROCESS_PID/PPID event
Ignore-this: 9c3ff130e72c48d8c94b0856434b7a84
The original EVENT_OSPROCESS_PID event contained the process pid and
the parent process pid. It turned out that this could not be supported
on windows (which has no real notion of parent process) so the event is
now split in two, pid and parent pid, and the latter is not generated
on windows.
  Technically this is an incompatible binary format change (removing a
field from an existing event) however the patch to make ghc generate
this event was only in for less than 24hrs before it got reverted (due
to not compiling on windows) so in reality it's fine.

darcs-hash:20110526223048-b0bbe-b591e1c77d8a20f21969b50ef380f2419cae2fc2